refactor: put send_request into httpcore
